requirements.txt
django==2.2.11
djangorestframework==3.11.0
django-cors-headers==3.2.1
python-decouple==3.3
google-auth==1.11.3
firebase_admin==4.0.0
pyfcm==1.4.7
gunicorn==20.0.4
pyjwt